More flexible promise support
The Lambda Node.js 8.10 runtime requires handler-returned promises
to be propper instances of the native Promise constructor. The
Node.js 6.10 runtime doesn't support handler-returned promises at
all. This change patches the runtime to use a more general
definition of a handler-returned promise that is also compatible
with the Node.js 6.10 runtime.